+++ Scripts/LayerExtensions.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+
+namespace Extension
+{
+    public static class LayerExtensions 
+    {
+        public static int LayerMaskToLayer(this LayerMask layerMask)
+        {
+            int layerNumber = 0;
+            int layer = layerMask.value;
+            while (layer > 0)
+            {
+                layer >>= 1;
+                layerNumber++;
+            }
+
+            return layerNumber - 1;
+        }
+    }
+}