+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## [1.0.0] - 31 May 2023
  - Initial Version
+## [1.0.1] - 4 June 2023
+ - Upgraded Rendere functions
+ - Added string and layers extensions